requirements.txt
anyio==3.6.2
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastapi==0.95.1
filelock==3.12.0
Flask==2.3.2
fonttools==4.39.4
h11==0.14.0
idna==3.4
ipykernel==6.23.0
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pip==23.1.2
platformdirs==3.5.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pydantic==1.10.7
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
unicorn==2.0.1.post1
urllib3==2.0.2
uvicorn==0.22.0
wcwidth==0.2.6
Werkzeug==2.3.4